US118248 Fixes and refactoring
