+++ components/tree-selector-node.js
@@ -1,6 +1,7 @@
 import '@brightspace-ui/core/components/icons/icon.js';
 import '@brightspace-ui/core/components/inputs/input-checkbox';
 
+import 'array-flat-polyfill';
 import { css, html, LitElement } from 'lit-element/lit-element.js';
 import { Localizer } from '../locales/localizer';
 import { RtlMixin } from '@brightspace-ui/core/mixins/rtl-mixin.js';
