+++ package.json
@@ -46,6 +46,7 @@
     "@brightspace-ui-labs/pagination": "^0.0.8",
     "@brightspace-ui/core": "^1.69.0",
     "@webcomponents/webcomponentsjs": "^2",
+    "array-flat-polyfill": "^1.0.1",
     "d2l-facet-filter-sort": "BrightspaceUI/facet-filter-sort#semver:^3",
     "highcharts": "^8.1.2",
     "lit-element": "^2",
